requirements.txt
pandas
numpy
google-cloud-storage
scikit-learn
pyyaml
imbalanced-learn
lightgbm
mlflow #for track of models
